requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
requests>=2.31.0
streamlit>=1.30.0
pandas>=2.1.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
rich>=13.7.0
plotly>=5.18.0
brotli
watchdog
cloudscraper
curl_cffi
playwright
